le gros push sa mère
projet bien avancé
avancement du 11/02/2022

+++ Assets/Scripts/Data/AttackData.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class AttackData : ScriptableObject
 {
     [SerializeField] private string _attackID = "ATTACK_ID";
-    [SerializeField] private float _damage = 10f;
+    [SerializeField] private int _damage = 10;
     [SerializeField] private float _actionTime = 5f;
     [SerializeField] private DatabaseManager.EAttackTypes _attackType = DatabaseManager.EAttackTypes.FIRE;
     [SerializeField] private DatabaseManager.ECombatEffects _combatEffect = DatabaseManager.ECombatEffects.NONE;
@@ -18,7 +18,7 @@ public class AttackData : ScriptableObject
      */
 
     public string AttackID => _attackID;
-    public float Damage => _damage;
+    public int Damage => _damage;
     public float ActionTime => _actionTime;
     public DatabaseManager.EAttackTypes AttackType => _attackType;
     public DatabaseManager.ECombatEffects CombatEffect => _combatEffect;
